Added the SFU Media Demo. Fixed copyrights. Moved RTP stuff back into an header file :)
